:memo: Establish Markdown anchors and new gitignore
Replaced all Markdown links with anchors at the end of each document.

Updated gitignore files for project and template from the Toptal URL.

Created the variables `scm_base_url` and `scm_link_url` for ease of use
within the template. For GitLab URLs, apart from the root of the
project, there is a additional segment with "/-/" in the URL, whereas
GitHub doesn't. `scm_link_url` was created in order to correctly render
both URLs.