:memo: Project README body changes
Updated minimal support version to Python 3.8 and modified sections to
include GitLab as a SCM option.

Added a "GitLab vs. GitHub features" section detailing differences in
the template when choosing one over the other.

Updated the "Input Variables" section with the latest Cookiecutter
variables.

Included a new topic in the "Makefile usage" section to refer to
linters.

Expanded the "What's next" section to guide the user with next steps for
setting up the project and reordered the additional libraries
recommendations.

Removed Release Drafter labels on the "Releases" sections.

Included a special acknowledgement to Roman Tezikov in the
"Acknowledgments" section for creating the original template in GitHub.
Included a link to Kedro's documentation detailing Kedro starters as a
reference and reading suggestion.

Added galactipy's reference via Kutt.