:memo: Segment Markdown anchors
Broke project and template READMEs Markdown anchors into groups to
facilitate updating in the future. For each section of the document, a
different code is used.

Updated the contents of the "Input variables" table to remove text
redundancy.

Added further instructions in the "What's next" section disclaiming the
restrictions of the template.