:pencil2: Replace most emojis with shortcodes
Replaced almost all emojis present in document bodies for their
shortcodes. The only remaining rendered emojis are present in
issue template comments to retain visual information for the user when
opening an issue in either GitHub or GitLab.

Fixed minor typo with an article title in the "Acknowledgements" section
of the project README.