:white_check_mark: Add semver unit tests
Created unit tests to check `validate_semver`, one for validating
seemingly invalid versions which are actaully valid and one for
validating seemingly valid versions which are actually invalid.

Semver regex worked as expected under the testing, however, it should be
noted those tests are not comprehensive of all possibilites a user might
input as a version.

Changelog: testing