:test_tube: Include tests for protected file/directory
Created failing tests intended to raise `PermissionError` when deleting
a path for which the user has no permission. First attempted using the
`Path.chmod` method, but it can only make a file/directory read-only,
which is fundamentally different from making it inacessible to remove.

Attempted using `shutil.chown`, but the test will raise a `LookupError`
should the user or group not be found on the system, making the test
harder to reproduce on different machines.

Possible implementations are via `shutil.chown` or by finding a way to
make the file immutable through the chattr command.

A xfail mark has been added to those tests, it is recommended to revisit
them in the future after researching a viable implementation under the
project's circumstances, possibly involving the CI pipeline.

Changelog: testing