:construction_worker: Enable Coveralls upload for project
Created GitLab CI job for uploading code coverage to Coveralls using
their Universal Coverage Reporter. Modified `test-py311` job to convert
`.coverage` file to XML in order to adhere to Coveralls format and
stored it in an artifact for use in the `coveralls-report` job.

The variable `COVERALLS_TOKEN` is stored in the GitLab repository for
access during pipeline execution to prevent leakage.

Changelog: ci