:white_check_mark: Add further tests to increase coverage
Created an additional test for `validate_username` in pre-gen hooks to
check against username length. Also included the actual statement in the
original function to raise the error when a username length is invalid.

Used pytest's `capsys` builtin fixture to test
`print_further_instructions` in post-gen hooks, although it could very
well be ignored since the original function's behaviour is to merely
print a string to the console.

With these additional tests, coverage reached 100%.

Changelog: testing