:heavy_minus_sign: Remove pytest-html from project dependencies
Since all coverage report for the project is now done through Coveralls
and the XML is generated from the `.coverage` file, there is no need for
the HTML report and pytest-html library for the project.

While this configuration will be kept for the template for now, once
Coveralls integration is built-in for GitHub and GitLab, they will also
be removed.

Changelog: dependencies