:heavy_plus_sign: Add pytest-click to template dependencies
pytest-click is a plugin that has not received an update since 2022, but
was developed for making it easier to write tests that interact with the
Click CLI tool package.

Since Click will be installed as a dependency for Typer if the user
selects `create_cli`, this plugin will also be installed should the user
want to write test for CLI commands.

Changelog: dependencies