:adhesive_bandage: Update importlib_metadata code
Since `galactipy` by default is only compatible with Python 3.8 and up,
there is no more need to separate cases for getting the version from
`pyproject.toml`.

Also, since the `name` key in the configuration is tied to
`project_name`, the `metadata` call was changed to correctly retrieve
the version (otherwise the CLI would always return "unknown" when
prompted with the `-v` flag).

Additionally, in order to adhere to Python project conventions,
`version` was changed to `__version__`.

Changelog: refactoring