:bug: Fix example color output
Fixed the color output for the CLI example. The issue was with the
`Color` class, whose values needed to be accessed via the `value`
attribute, otherwise the `color` variable in `main` would end up being
`Color.red` instead of the string `red`, for instance, making Typer
unable to render colors.

Changelog: bugfix