:bug: Fix behaviour for file removals
The boolean variables in `cookiecutter.json` need to be invoked in
`post_gen_project.py` as strings to avoid errors when testing galactipy.
However, as strings, the variables `CREATE_CLI`, `CREATE_DOCKER` and
`CREATE_DOCS` will always be interpreted as True, even if the user
chooses a false value during the prompt.

Since the `bool` function will convert the string "False" as True since
it is non-empty, and using `ast.literal_eval` will raise errors when
testing, the chosen solution was to evaluate the string of the
cookiecutter variable against the "True" string.

Changelog: bug