:rotating_light: Select pydocstyle rules for project
Ignored some of the rules in several levels.

For anything inside "tests" folders, ignored the lack of docstrings for
modules, classes and functions.

Since the project uses numpydoc, ignored violations of one-line
docstrings as numpydoc always implements a blank line.

For the hooks, ignored the lack of docstring for the `main` function.

For the `something` function in the template testing module, ignored the
rule for imperative mode.

Changelog: fix