:rotating_light: Fix minor linter warnings for template
Set Ruff to ignore docstring rules in the `__main__.py` file, wrapped
comment lines to adhere to the 79-character limit and added a period at
the end of the `__init__.py` module docstring.

Ignored the ARG001 rule in `__main__.py` for the `print_version`
argument, since it is covered by the callback of Typer.

Changelog: fix