:wrench: Improve compilation task syntax
Changed the Invoke implementations of `lint_all` and `cleanup` to
explicitly call their tasks within the function, instead of pre-tasks
using the decorator.

This is a better option to identify tasks which are simply collections
of others tasks and differentiate those from tasks that actually have
other tasks as their dependencies.

Changelog: build