:bug: Fix Python path for tasks
If a developer is to install Poetry and then use it to install
dependencies in the environment, they will most likely run
`poetry_download` before anything else, and if there is no active venv,
the command will fail, as no Python executable will be found.

A fix is provided so that if no venv is active, it will try to fetch the
available executable in the system.

Changelog: bugfix