:memo: Update READMEs with Invoke references
With the complete implementation of Invoke, project and template
READMEs, including supporting files such as `CONTRIBUTING.md` and the
`Docker` folder `README.md` have been updated, including a revision of
the Invoke commands section to simplify the categories.

Kudos to the Pelican team being acknowledged with this commit after the
support given to understand best practices for laying out the `tasks.py`
file.

Changelog: documentation