:adhesive_bandage: Fix Docker Invoke commands
The Docker commands were fixed as they were not fully functional,
starting with the required PATH of the `build` command.

Additionally, a workaround was needed to make iterable arguments to
work, as for some reason Invoke will not allow a default argument in
this case. Thus, the `DEFAULT_DOCKER_IMAGE` is called in the function
body, ensuring that the command will run with a single or multiple
images specified by the user or with no argument at all.

Changelog: bugfix