:memo: Update READMEs with VS Code settings
In addition to listing the predefined settings as a feature, the
documentation is also updated to reflect the use of the
`docstring_length` parameter, removes the boilerplate Todo Tree
configuration and adds an article for reference on how to use TODOs.

Additionally, a few items in the "What's Next" section for the template
are formatted with to-do items, to be shown up when the user creates the
project and uses Todo Tree.

Changelog: documentation