:children_crossing: Specify Poetry publishing exceptions
Refined the `publish` Invoke command for the template to warn the user
about the two most common errors when attempting to publish the package
to PyPI.

As Python exceptions in these cases are only available as STDOUT and
thus unreachable by code, the behaviour of the failed command is to
display the complete traceback, followed by the custom warning message.

Also reworked the command's flags to prevent an unwanted behaviour with
Poetry by not recognising the default PyPI repo as a publishable.

Changelog: enhancement