:construction_worker: Move SemVer regex to pipeline variables
As a way to make the GitLab CI configuration more organised, the regex
expressions for evaluating a full release version and one containing
optional pre-release and build parts have been moved.

Also removed the negative lookahead for end of string in the non-release
version, as these are not supported by RE2. Thus, an additional
condition is needed to validate that the tag is not a full version.

Changelog: template-config