:truck: Separate test and production publishing for GitHub Actions
As the `on` keyword is defined at the workflow level, publishing to PyPI
and test PyPI have been separated. The syntax for publishing to the test
repository includes the additional parts for pre-release and build as
close as it could get given GitHub glob restrictions, while enforcing
that full release versions will be ignored for this worklow.

Still pending on testing and confirming that escaping "+" with a
backlash works in GitHub actions.

Changelog: enhancement