:construction_worker: Refine rules for automatic releases
Changed the CI configuration so Python setup is not used for the release
stage. A future usage of a Docker image already containing Invoke and
Poetry is suggested.

The changelog commit message is changed so the URL is correctly formed
and a rule is added so the CI blocks any pipelines started from bots, as
the only change in the changelog commit is the Markdown file.

Changelog: ci