:construction_worker: Refine rules GitHub workflows
As the `on` keyword works under "or" logic when multiple rules are
added, the rule for publishing to PyPI in the production environment is
adjusted so it is always started from the manual run, but only
effectively publishes the package if ran from a tag.

This leaves room for a non-release tag to be accidentally published, but
given GitHub Actions limitations, the usage of workflow dispatch
diminishes this risk.

Additionally, the release drafter's rules are also altered so it only
runs on release tags.

Changelog: template-config