:construction_worker: Create job to update github-mirror branch
Added a job for the template to automatically update the `github-mirror`
branch when a changelog commit is made by Galactipy Bot. A file CI/CD
variable is defined for the project with the options to be used for
creating the merge request.

If possible, a future development should address the "ad hoc" nature of
the merge action itself, as it relies on the previous action to store
the MR's ID within the job script.

Changelog: ci