:wrench: Fix existing Docker tasks
Modified so the `DEFAULT_DOCKER_REPOSITORY` variable for Invoke contains
the full path for Docker Hub.

The `docker_build` and `docker_remove` tasks were modified so the user
provides only tags to the command and the task itself will handle
processing the correct Docker command. The user can now also provide an
alternate repository for building and removing.

Changelog: template-config