:sparkles: Add GitHub Docker push workflow
Used the official Docker GitHub Action to structure a functional
workflow to push images to Docker Hub. Currently, the workflow functions
under a workflow dispatch for the tags for input, but only pushes a
single "latest" tag. Tests should be done with generated templates to
refine the workflow's behaviour

Changelog: feature