:construction_worker: Update project CI to use custom Docker images
Changed the CI configuration to use the `galactipy/docker` images with
Invoke and Poetry pre-installed, removing the need for the `.python_setup`
job template and only moving the project installation command to the
`.default_scripts` job template.

With the custom images being used to speed up testing, the `default`
section of the CI has removed the Python image, and the Coverall upload
now uses the Alpine image in order to reduce compute minutes.

Changelog: ci