:wrench: Add exception message for nonexistent Invoke installation
Included a try/except block to check if Invoke is available before
running `tasks.py`, which raises a message when not found.

While good practice, in reality this message will only be displayed if a
developer wrongfully attempts to run `tasks.py` directly via Python, as
a nonexistent Invoke installation will not even be recognised by the
shell before running.

Changelog: project-config