:building_construction: Revamp path resolution for project tasks
In order to make the move to the new custom Docker images for the
project CI, the `tasks.py` has been reworked to ensure Invoke knows how
to handle virtual environments on a more general level.

The steps for resolving virtual environment path, as well as the Python
and Poetry paths, has been moved to a `utils` package and its logic
encapsulated in two functions: `get_venv_bin` and `get_commands`, while
the API has been reworked to work on more cases than the previous code.

One of the problems faced in moving to the new images is that `tasks.py`
assumes work is being done on a virtual environment, which was created
as a step in the CI jobs involving Python. However, with the new images
providing exactly what is needed for minimal testing, creating venvs
inside them feels redundant.

Instead, the `utils` package now ensures either a path to a bin
directory in which Python package executables should be found or None is
returned, prioritising an active venv, the virtualenvwrapper home
directory and a Poetry venv, in this order, are searched.

By establishing the `get_venv_bin` function, we ensure the paths are set
every time Invoke is called and, if there is none available when Poetry
creates its lock file, the function is called a second time to ensure
the automatically created venv is used.

Changelog: build