:construction_worker: Refine project CI rules and structure
Simplified the project's CI syntax to leverage template jobs to define
rules for tests. As essentially they only need to run when developing
new features and on the master branch for publishing coverage to
Coveralls, the rules are better detailed to ensure the stage is only run
when needed.

Also updated the rule to bypass pipelines if created by a bot user,
considering now the project's development on a dedicated group.

Changelog: ci