:wrench: Include GitHub workflow removal for Docker
With the new feature, the post-gen hooks were updated to reflect the
possibility of a user choosing the GitHub platform without Docker, in
which case only the `docker.yml` workflow should be removed.

In addition to changing the hooks and updating tests related to Docker
removal, which included separating tests for GitLab and GitHub in this
scenario, the `template_tree` fixture was updated to reflect the current
state of the template structure.

Closes #104, #105

Changelog: hooks