:memo: Update READMEs with BDD functionality
A small explanation to BDD and `pytest-bdd` is added to the READMEs, and
a custom badge for when the user chooses to use BDD is added to the
template `README.md`.

Additionally, more information was given about the available badges and
divided the template badges into four distinct sections.

Closes #124, #127

Changelog: documentation