:construction_worker: Replace Coveralls with Codacy
Several changes were made to the CI file, the main one being the
replacement of Coveralls as the coverage tool in favour of Codacy, using
a similar approach to the previous job for Coveralls.

Additionally, the CI rules are revised and refined to ensure proper
running of the test jobs, including the fact that previously no merge
request pipelines were being run.

Finally, the `changelog-commit` job is changed to move the payload to a
CI/CD variable and being provided as cURL data, removing the need for
the commit message variable in the file.

Changelog: ci