:memo: Add Codacy badges for the project
With the choice of Codacy as the project's QA platform, the Coveralls
badge has been replaced, while an additional badge for code quality has
been added.

The sections for the project's badges have also been rearranged to match
the template's model of project info, contributions, guidelines and
tools.

Closes #135

Changelog: documentation