:sparkles: Add option between code coverage and code analysis
Created a `coverage_service` Cookiecutter variable to allow users to
choose between code coverage with Coveralls and code analysis with
Codacy.

The option for these two services came to be as Coveralls offers a basic
interface aimed only at test coverage reports, which is sufficient if
the user is just looking for a simpler option to manage the project,
while Codacy offers more features besides code coverage.

Changelog: feature