:sparkles: Set up GitLab coverage uploads
Modified the CI workflow for templates by making the minimal Python
version test job generate coverage files (this is the opposite of what
happens to the Galactipy CI) and then add a second job in the for loop
to upload coverage to the chosen code quality service.

The `coverage` Invoke task is also implemented for the template,
following the same pattern as the project task.

Closes #139, #140

Changelog: feature