:sparkles: Implement GitHub coverage action
Added steps to the `test` GitHub workflow to include coverage generation
and upload to the selected coverage service. Unlike the Galactipy
project, generated templates pin the minimal Python version as the
version to use for uploads to the service.

Also updated the `docker` workflow to explicitly state the need for
storing the Docker PAT on the repository secrets.

Closes #137, #138

Changelog: feature