:sparkles: Split test and coverage uploads for GitHub templates
Since the `if` rules for a step are processed before `matrix` in GitHub
Actions, the coverage for generated projects was being uploaded for all
Python versions. To address this, a reusable workflow has been created
which allows the specification of the Python version to run tests on and
if the coverage data should be uploaded to the code quality service.

With this, the `test` workflow is rewritten so the minimal Python
version tests and uploads coverage, while all others only test the
repository, mirroring the behaviour of GitLab templates.

Changelog: enhancement