:building_construction: Reorganise template with new CLI structure
Reworked the template structure to adapt generated projects to the new
logic of creating a CLI by default, while the "bare repo" option remains
the alternative. All conditional statements that used the `create_cli`
variable have been inverted for the new `bare_repo` variable.

The most striking changes made to the template is the removal of all
example code in the `example.py` and `__main__.py` modules, aiming to
facilitate actual usage of Galactipy as a starter template for projects.

In place of the previous CLI example with the `--name` and `--color`
options, a bare structure for a CLI package using Typer is added. In its
initial state, only a command with the `--version` argument is provided
inside the `root_command.py` module, while a `commands` directory is
included for development of CLI applications using Typer's directive of
one command per file.

The `pyproject.toml` is updated with the new entry point for the CLI.

Changelog: enhancement