:wrench: Replace issue templates for project
Replaced the original `python-package-template` templates for issues,
which were kept basically the same ever since Galactipy started being
developed, with a more thought-out categorisation using "requests".

This model is aimed at giving users unrelated to the project development
clearer guidelines as well as allowing developers to more easily
navigate through the sort of request incoming from outside the project.

While originally the idea was to only have a single template for issues,
focusing entirely on Merge Request templates to implement GitLab
philosophy, this requires at least a branch to be associated with the
MR, which would impose a barrier for those who are not already
collaborators of the project.

Changelog: project-config