:wrench: Add template for internal issues
Added a fourth issue template exclusively for dealing with the internal
unravelling of Merge Requests. Ideally, under the GitLab philosophy, the
development team will first create MRs to define the "what" of
development, while they can be linked to multiple issues and tasks which
will be responsible to define the "how" of implementation.

Tasks are to be used as acceptance criteria for issues created with this
model.

Changelog: project-config