:pencil: Rebuild contributing guidelines
The entire `CONTRIBUTING.md` file has been changed to provide a
comprehensive guide on open source contribution. Through a number of
references, several sections were included to touch on project policies,
contributing guidelines and orientation for common users. This model
should be used to also provide a more comprehensive `CONTRIBUTING.md`
file for the template in a future development.

To mark the evolution of Galactipy as a truly open source software, the
copyright reference has been altered in `pyproject.toml`.

Also changed Galactipy's versioning logic to EffVer as SemVer does not
make sense in the case of Cookiecutter templates.

Closes #24, #148, #150, #151, #152, #161, #179, #195, #196, #197, #198, #209

Changelog: policies