:wrench: Replace MR templates for project
Replaced the original `python-package-template` Merge Request template
with multiple categories to facilitate developer understanding of
changes in larger scale projects.

While there are similar descriptions to the issue templates, the main
focus of these files is to standardise a checklist of items which
contributors should follow to avoid any missing parts when submitting
changes.

This model, based on the GitLab philosophy of "proposals", should be
carried to the template in a future development, with specific
categories relating to CLI applications.

Closes #49

Changelog: policies