:recycle: Encapsulate removal flags in dataclass
To implement the feature flags based on file or directory names, a new
function for retrieving the list of files will be necessary, which will
violate the maximum number of arguments allowed by Ruff.

The post-gen hook file is changed to move the flags to a dataclass and
reduce the number of arguments.

Changelog: hooks