:technologist: Add feature flags functionality
A hidden `__debug` Cookiecutter variable was added. When this value is
set to true, it will trigger the post-gen hooks to keep any directories
and filenames ending in `_feature` after project generation.

Unlike other removals, which are all retrieved in
`_get_files_to_delete`, the feature files are only evaluated after all
other removals have been applied. This prevents a double count of files
to delete (e.g., when adding a `_feature` file under `docker` and
choosing to not render a project with Docker).

This will enable Galactipy developers to iteratively add changes to the
repository without having to manually set files to delete nor provide
incomplete features to users.

Changelog: hooks