:sparkles: Create demo TUI application
Provided a simple application using Textual to be placed in the `tui`
directory. This application uses `art` to generate a stylish startup
screen with an initial configuration so users can build up from there.

While this implementation breaks the design principle of avoiding extra
content that will not carry to user projects, it does so with minimal
intervention: users just need to remove two lines of code and the `.tcss`
file to have a clean slate for their projects.

Closes #211

Changelog: feature