:art: Include TUI launch command for the CLI
Depending on the application type chosen, the TUI will either be
launched from the "raw" CLI command or with the "launch" command. Jinja
conditionals were added to the root command file to reflect this
feature, along with a separate `launch.py` file if the "hybrid" option
is chosen at startup.

Further development is required to address the need for file removals,
considering the interaction between the TUI and CLI sides of the
application.

Changelog: enhancement