:wrench: Specify TUI-related files for removal
Added a `APP_TYPE` constant in post-gen hooks to control the behaviour
of the `_get_tui_specific_files` function. When the project is generated
either as an integrated TUI or CLI-only, there is no need to create a
`launch` command for the program, while in CLI-only or bare repo
options, there is no need for anything inside the `tui` directory.

Changelog: hooks