:wrench: Create dedicated function for BDD file removals
With the increased complexity to deal with the interaction between
application types and the usage of BDD for generated templates, a new
function was added to retrieve all files related to BDD specifically.

As a side effect, this reduces the complexity of the
`_get_cli_specific_files` function, which could be even deprecated. It
was chosen not to in the name of consistency.

Changelog: hooks