:pencil: Add work item tracking policy for project
With the release of GitLab Statuses as a feature, the project has
implemented custom categories to better track progress for issues and
tasks. A section was added to better explain each status and its usage.

This effectively removed the need for the `needs` labels, so sections
referring to them were either removed, toned down or adapted to refer to
statuses instead.

References to the `seeking-contributors` labels have also changed to
reflect the renamed `seeking-contributors::delivery` label, also a
consequence of implementing statuses for the project.

We took the opportunity to convert some of the changed sections to SemBr
notation as well.

Changelog: policies