:wrench: Add sample unit test for TUI main window
Included an example test based on the Textual documentation to validate
that Textual terminates the application correctly, as a starting point
for users to build their own tests for their TUIs. This included
`pytest-asyncio` as a dev dependency for integrated and hybrid projects.

A scenario was created to implement the unit test using BDD, but the
code is not functional at this point due to incompatibilities between
`pytest-bdd` and `pytest-asyncio`, which must be addressed before
officialising the TUI feature.

As a minor fix, the fixture for the version string was corrected to
reflect the removal of a whitespace in the application.

Changelog: template-config