:white_check_mark: Rework file removal tests
With the changes in post-gen hooks and the increased complexity, an
extensive revision of the `removal_tree` fixture and the tests for file
removals was required. In order to streamline the usage of the removal
tree and reduce the need for individual references depending on the test
being run, the fixture now returns a complete work tree of paths for
everything that is being tested.

A future development should refactor this fixture, probably along with
the `bulk_file_creation` function, to further streamline the codebase.

Changelog: testing