:green_heart: Ingore rule for top-level import in post-gen hooks
With the update to the latest version of Ruff, a linter error prevented
the CI from passing as it required all imports to be defined at the top
of the module. This was suppressed instead of fixed due to possible
errors when users generated a template without the module in question
being available in their environment.

Changelog: bugfix