:wrench: Revamp Dockerfile structure for generated projects
Reworked the `Dockerfile` for the template, improving layering, image
size and contemplating the three different options for applications from
the `app_type` Cookiecutter variable.

The image now employs multi-stage building and leverages
`poetry-plugin-bundle` to create a virtual environment and provide only
the necessary binaries to run the application. For integrated and hybrid
TUIs, this defaults to launching the interface. For CLI apps, an
`ENTRYPOINT` is defined to always run the top level command.

For bare repositories, a token `CMD` instruction is given to raise a
message informing of non-existing code to run, since there are no files
besides `__init__.py`.

Closes #231, #232, #291, #292, #293

Changelog: template-config