:construction_worker: Update project CI with new task nomenclature
Changed the Invoke task commands to reflect the changes in task names.

Additionally, made minor improvements to the file to remove the need for
the local `GALACTIPY_BOT_TOKEN` variable (now relying on the group
access token) and ensure that the CI will run if any files under `tests`
are changed, not only those directly under its root.

Changelog: ci