:wrench: Rewrite additional instructions output
Improved the contents of the `print_further_instructions` function to
reflect the changes in Invoke tasks. Additionally, included a
conditional to check the existence of Poetry in the user's system,
printing a similar warning to that of Invoke.

Instead of using `find_spec` for Poetry and Invoke, `which` has been
used as the former is preferred to locate actual Python packages, and
not command-line programs.

The tests and fixtures to validate the rewritten, improving object
mocking, but also marking one of the tests with the `xfail` mark, as the
simulated Rich `Console` object defaults to a 80-character width and
attempting to force the `galactipy_poetry_instructions` line break is
not sufficient to make both strings match.

Changelog: build