:rewind: Revert GitLab Changelog modifications
After merging the latest branch to master, the tag pipeline suceeded,
but provided no content to the changelog. When using `glab` locally to
generate any changelog, even to past tags, an error 422 is raised by the
API with the message:

```
Failed to generate the changelog: Failed to parse the template: Extra
input after last repetition at line 14 char 1.
```

Since changing the file locally and running the `glab` command with an
explicit `--config-file` option yielded the same result, the best option
is to revert those changes to check:

1. That the `glab` command will always point to the upstream
   `changelog_config.yml` file;
2. If the issue is with changes in the Changelog API from GitLab.