:sparkles: Implement configuration commands in the CLI
Added basic CRUD operations for the CLI application to directly modify
the settings files, which includes a command to handle arrays in the
TOML settings file.

As Typer does not support type annotation unions, `ast.literal_eval` was
used to enable coalescing of strings to common Python types.

Closes #206

Changelog: feature