:white_check_mark: Add config tests
Added the unit tests for all commands related to configuration
management, both for BDD and traditional testing, which also includes
validating the behaviour of helper functions to instatiate and resolve
`AppManager` objects.

In general, fixtures to provide preset configuration states were added,
with the need for mocking the `get_default_config` function to prevent
modifications to spread to a developer's filesystem.

For BDD, applied step parsing extensively to streamline code building.
This also required creating a parser to convert strings to boolean
objects for convenience when binary options are used in the CLI.

Closes #305

Changelog: testing