:sparkles: Add option for GitLab Premium/Ultimate
Updated `cookiecutter.json` with an additional option for the
`scm_platform` variable so the user can specify GitLab Premium or GitLab
Ultimate as their tier for version control.

Currently, the option is merely aesthetic and has no effect whatsoever
in the template, but with the future developments of #199, #233, #290
and the intention to implement %13, differentiating this option will be
relevant.

Closes #246, #247

Changelog: enhancement