:speech_balloon: Refine configuration variables and values
Took the opportunity to revisit the Cookiecutter configuration, mainly
replacing the term "username" with "namespace", which is clearer and
less prone to confusion.

Also removed the last remaining mention of the `version` variable in the
prompts, updated the input variables table in the project README and
formatted the tables in the file.

Changelog: fix