:wrench: Refine template linting behaviour
During development of Nebulog, many issues arose from the overly strict
rules for MyPy extending to the test files. Overhauled MyPy
configuration for the template, keeping strictness for code and allowing
more freedom with type checks for the test files.

Additionally, ignored rules for the Ruff linter in test files and
`tasks.py` for the same reasons.

Changelog: template-config