:sparkles: Implement logging capabilities for the template
Added Nebulog to the template dependencies along with a `logging` module
for opinionated configuration for loggers. This involves a preconfigured
format for log calls stored in the log file, with paddings for the level
and the message.

The padding for the message was set to 60 as we don't expect log
messages to be overly long, while also setting a value that would not
make reading of columns further to the right more difficult.

Padding for the file location is obtained by getting all module names
for the project being built with the template and setting to the longest
filename.

A top-level `setup_app_logging` is defined to preconfigure the default
Nebulog logger, along with a traditional `.log` file and a serialised
`.json` log.

Closes #339

Changelog: feature