:art: Add helper code for CLI value type resolution
Created a `helpers` package for the CLI to abstract how the program will
handle input and output for config operations.

For the `get` command, the `pretty_print_setting` evaluates whether the
returned configuration should be cast into a dict or a list. For the
`set` and `extend` commands, a simple API is defined for usage with the
annotated Typer argument/option with a parser.

As mentioned in the module docstring, Typer does not support union type
operators, so users have to either design their own solutions or use
`pydantic-typer` as a plugin to integrate Pydantic validation into
Typer. We chose to implement a separate boilerplate API to avoid
declaring Pydantic and `pydantic-typer` as dependencies before resolving
orbittings#7.

Changelog: enhacement