:art: Simplify backend API
Made improvements to code related to CLI type converters and logging
formatters. Scrapped unnecessary code which was to be used with the
shell logger, which is now handled directly by Nebulog, streamlined the
API for the `BasicConverter` class while also removing the
`parse_converter` function, as Typer allows receiving a class for its
`parser` argument.

Changelog: enhancement