:art: Improve CLI commands API
Made some minor improvements to the structure and behaviour of code
involving the CLI application commands. Removed the calls to
`setup_app_logging` on the configuration commands as the context from
the root command will ensure logging is stored in the files.

Also added a more detailed help message for the set and extend commands
to point the need for double dashes when passing a negative number.

Changelog: enhancement