:children_crossing: Add scenarios to test TUI launch
One of the reasons coverage was not reaching 100% on hybrid applications
was the lack of logging formatter coverage. By mocking the terminal
interface and adding a single test to check the `launch` command
execution, the target coverage rate is achieved.

The same test has also been added for integrated TUI applications to
keep standardisation, but it wasn't necessary for coverage per se.

Closes #335

Changelog: enhancement