:sparkles: Overhaul Git trailers for the template
Replaced the original `python-package-template` Git trailers for release
notes with an extensive and renewed list of categories tweaked
specifically for CLI/TUI applications, since they are the main feature
of Galactipy templates.

Our intent was for developers to be able to better convey changes in
more specific domains without being overly technical with terms like
"backend" or "entry points", instead using categories to communicate
changes as groups ranking in relevance to the end user.

On the other hand, several Git trailers have been defined so developers
have less trouble remembering which names they should use to ensure
correct rendering in the final release. This will help different
projects to use those trailers without requiring their developers to
memorise terms that may not be relevant for their specific project.

This approach not only allows generated projects to require less work on
the development team to communicate changes in a meaningful way to
users, but also lifts some of the burden on the developers,
incentivising more modular development increments, in line with
Galactipy's philosophy.

Closes #357

Changelog: template-config